datasource/realtime/requirements.txt
numpy
pandas
kafka-python